Add files via upload
Hypothesis testing is a fundamental aspect of inferential statistics. It helps in making decisions or inferences about population parameters based on sample data. The idea is to test whether there is enough evidence in a sample of data to infer that a certain condition is true for the entire population.